changed logic to handle possible foreign key violation before performing DML calls

+++ spring-jpa/src/main/java/com/example/springjpa/customer/service/CustomerServiceImpl.java
@@ -4,19 +4,23 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.example.springjpa.common.Status;
 import com.example.springjpa.common.exception.DeletingCustomerWithExistingOrdersException;
 import com.example.springjpa.customer.dao.CustomerDao;
 import com.example.springjpa.customer.domain.Customer;
+import com.example.springjpa.order.service.OrderService;
 
 @Service
 public class CustomerServiceImpl implements CustomerService {
 	
 	@Autowired
 	CustomerDao customerDao;
+	
+	@Autowired
+	OrderService orderService;
+	
 
 	@Override
 	public List<Customer> getCustomers() {
@@ -49,15 +53,17 @@ public class CustomerServiceImpl implements CustomerService {
 	@Override
 	public Status deleteCustomer(Integer customerId) {
 		
-		try {
-			
+		boolean emptyOrders = orderService.getOrdersByCustomerId(customerId).isEmpty();
+		
+		if( emptyOrders ) {
+		
 			customerDao.deleteById(customerId);
 			
 			return Status.SUCCESS;
 			
-		} catch (DataIntegrityViolationException e) {
-
-			throw new DeletingCustomerWithExistingOrdersException(e);
+		} else {
+			
+			throw new DeletingCustomerWithExistingOrdersException();
 		}
 	}
